# contrat/openapi.yaml
openapi: 3.1.0
info:
  title: Plateforme de gestion des réclamations — API
  version: 1.0.0
  summary: Contrat de l'API du MVP (Makor Telecoms · Solution 1)
  description: |
    Contrat de l'API, écrit avant le code (étape 5). Le backend NestJS de l'étape 7 et le frontend
    de l'étape 8 s'y conforment ; les réponses de l'API seront validées contre ce fichier par les tests.

    ## Espaces

    | Préfixe | Qui | Authentification | La banque est connue par |
    |---|---|---|---|
    | `/public` | Client final, sans compte | aucune | le code du point de dépôt ou le jeton de suivi |
    | `/client` | Client final après code OTP | `jetonClient` (30 min) | le jeton |
    | `/auth` | Personnel | mot de passe + TOTP | — |
    | `/banque` | Agent, superviseur, Admin Entreprise | `jetonPersonnel` (15 min) | le jeton |
    | `/plateforme` | Super Admin | `jetonPersonnel` | — (toutes les banques, métadonnées seulement) |

    La banque n'est jamais un paramètre libre de la requête : un agent de la banque A qui demande
    un ticket de la banque B reçoit 404.

    ## Conventions

    - Chemins et champs en français, comme le code et le modèle de données.
    - Dates au format ISO 8601 en UTC (`2026-09-25T10:00:00Z`) ; l'interface les affiche dans le
      fuseau de la banque (`fuseauHoraire` de `/banque/parametres`).
    - Identifiants UUID.
    - Listes paginées : `page` (à partir de 1) et `parPage` (100 au plus) ; réponse
      `{ donnees, pagination: { page, parPage, total } }`.
...
servers:
...
paths:
...
  /sante:
    get:
      tags: [Santé]
      operationId: lireSante
      summary: État de l'API, de la base et de Redis
      x-roles: [PUBLIC]
      security: []
      responses:
        '200':
          description: Service opérationnel
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sante' }
        '429': { $ref: '#/components/responses/TropDeRequetes' }
        '503':
          description: Une dépendance ne répond pas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Sante' }

  # ===========================================================================
  #  Public : dépôt et suivi
  # ===========================================================================
...
  /banque/parametres/logo:
    put:
      tags: [Paramétrage]
      operationId: televerserLogo
      summary: Logo du portail (PNG, SVG ou WebP, 1 Mo au plus)
      x-roles: [ADMIN_ENTREPRISE]
      security:
        - jetonPersonnel: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [logo]
              properties:
                logo: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/Parametres' }
        '400': { $ref: '#/components/responses/Invalide' }
        '401': { $ref: '#/components/responses/NonAuthentifie' }
        '403': { $ref: '#/components/responses/Interdit' }
        '413': { $ref: '#/components/responses/FichierTropVolumineux' }
        '415': { $ref: '#/components/responses/TypeNonSupporte' }
...
components:
  securitySchemes:
    jetonPersonnel:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton d'accès du personnel (15 minutes), obtenu après mot de passe + code TOTP.
    jetonClient:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session client (30 minutes) obtenue avec le code OTP ; limitée à ce client et à sa banque.
    cookieRafraichissement:
      type: apiKey
      in: cookie
      name: rt
      description: Refresh token (httpOnly, Secure, SameSite=Strict), remplacé à chaque usage.

  # ---------------------------------------------------------------------------
...
  responses:
    Invalide:
      description: Requête mal formée ou champ invalide (code VALIDATION, erreurs par champ)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Probleme' }
    NonAuthentifie:
      description: Jeton absent, expiré ou invalide ; identifiants ou code incorrects
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Probleme' }
    Interdit:
      description: Rôle insuffisant, ou banque suspendue (code BANQUE_SUSPENDUE, pour toute opération du personnel de cette banque)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Probleme' }
...
    TropDeRequetes:
      description: Limite de débit atteinte
      headers:
        Retry-After:
          description: Secondes avant de réessayer
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Probleme' }
    FichierTropVolumineux:
      description: Fichier au-delà de 5 Mo (1 Mo pour un logo)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Probleme' }
    TypeNonSupporte:
      description: Type de fichier refusé (JPEG, PNG, WebP ou PDF attendus)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Probleme' }
...
    Parametres:
      description: Paramètres de la banque
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ParametresBanque' }
...
  schemas:
...
    CodeErreur:
      type: string
      description: Code stable d'une erreur, à utiliser par les interfaces (le titre peut changer)
      enum:
        # Requête et accès
        - VALIDATION
        - NON_AUTHENTIFIE
        - JETON_INVALIDE
        - INTERDIT
        - INTROUVABLE
        - TROP_DE_REQUETES
        - CONFLIT_IDEMPOTENCE
        # Authentification
        - IDENTIFIANTS_INVALIDES
        - COMPTE_VERROUILLE
        - CODE_TOTP_INVALIDE
        - MOT_DE_PASSE_TROP_FAIBLE
        - CODE_OTP_INVALIDE
        - CODE_OTP_EXPIRE
        - TROP_DE_TENTATIVES
        # Machine d'états (étape 4)
        - TRANSITION_INTERDITE
        - ACTEUR_NON_AUTORISE
        - AUCUN_AGENT_ASSIGNE
        - DELAI_DE_CONTESTATION_DEPASSE
        - CLOTURE_AUTOMATIQUE_PREMATUREE
        # Dépôt et messages
        - BANQUE_SUSPENDUE
        - POINT_DE_DEPOT_INACTIF
        - CATEGORIE_INVALIDE
        - CONTACT_REQUIS
        - CONTACT_INVALIDE
        - DESCRIPTION_REQUISE
        - CONSENTEMENT_REQUIS
        - ANTI_ROBOT_REFUSE
        - MESSAGE_VIDE
        - PRECISION_REQUISE
        - AGENT_INVALIDE
        # Fichiers
        - FICHIER_TROP_VOLUMINEUX
        - TYPE_DE_FICHIER_NON_SUPPORTE
        - TROP_DE_FICHIERS
        # Paramétrage et plateforme
        - PLAFOND_AGENTS_ATTEINT
        - EMAIL_DEJA_UTILISE
        - NOM_DEJA_UTILISE
        - CODE_DEJA_UTILISE
        - PREFIXE_DEJA_UTILISE
        - SLUG_DEJA_UTILISE
        - QR_CODE_SANS_AGENCE
        - SUPERVISEUR_INVALIDE
        - INVITATION_DEJA_ACCEPTEE
        - JOUR_FERIE_EXISTANT
        # Serveur
        - ERREUR_INTERNE

    # ---- Communs -----------------------------------------------------------------
    Probleme:
      type: object
      description: Erreur au format RFC 9457
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference, examples: ['/erreurs/transition-interdite'] }
        title: { type: string, examples: ['Action impossible dans le statut actuel'] }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string, examples: ['RESOUDRE est impossible depuis le statut OUVERTE'] }
        instance: { type: string, examples: ['/api/v1/banque/reclamations/0199.../resolution'] }
        code: { $ref: '#/components/schemas/CodeErreur' }
        erreurs:
          type: array
          description: Détail par champ (code VALIDATION)
          items: { $ref: '#/components/schemas/ErreurChamp' }
    ErreurChamp:
      type: object
      required: [champ, message]
      properties:
        champ: { type: string, examples: ['telephone'] }
        message: { type: string, examples: ['Numéro de téléphone invalide'] }
...
    Sante:
      type: object
      required: [statut, base, redis, version]
      properties:
        statut: { type: string, enum: [ok, degrade] }
        base: { type: string, enum: [ok, indisponible] }
        redis: { type: string, enum: [ok, indisponible] }
        version: { type: string, examples: ['1.0.0'] }
...
    Couleur:
      type: [string, 'null']
      pattern: '^#[0-9A-Fa-f]{6}$'
      examples: ['#0B5FFF']
...
    ParametresBanque:
      type: object
      required: [nom, slug, prefixeTickets, fuseauHoraire, seuilAlerteSlaPourcent, delaiClotureAutoJours,
        smsChaqueChangementStatut, couleurPrimaire, couleurSecondaire, logoUrl, emailContact, plan, consommation]
      properties:
        nom: { type: string }
        slug: { type: string }
        prefixeTickets: { type: string }
        fuseauHoraire: { type: string }
        seuilAlerteSlaPourcent: { type: integer, minimum: 1, maximum: 99 }
        delaiClotureAutoJours: { type: integer, minimum: 1, maximum: 60 }
        smsChaqueChangementStatut: { type: boolean }
        couleurPrimaire: { $ref: '#/components/schemas/Couleur' }
        couleurSecondaire: { $ref: '#/components/schemas/Couleur' }
        logoUrl: { type: [string, 'null'], format: uri-reference }
        emailContact: { type: [string, 'null'], format: email }
        plan:
          type: object
          required: [nom, plafondAgents, plafondTicketsMois]
          properties:
            nom: { type: string }
            plafondAgents: { type: [integer, 'null'], description: "Vide = illimité" }
            plafondTicketsMois: { type: [integer, 'null'], description: "Vide = illimité" }
        consommation:
          type: object
          required: [agents, ticketsCeMois]
          properties:
            agents: { type: integer, description: "Agents et superviseurs non désactivés" }
            ticketsCeMois: { type: integer }
...